move cli tools e2e tests
