+++ cli_tools_e2e_test/gce_image_import_export_tests/main.go
