+++ cli_tools_e2e_test/gce_ovf_import_tests/compute/instance.go
