+++ cli_tools_e2e_test/gce_ovf_import_tests/go.mod
@@ -1,0 +1,15 @@
+module github.com/GoogleCloudPlatform/compute-image-tools/gce_ovf_import_tests
+
+go 1.13
+
+require (
+	github.com/GoogleCloudPlatform/compute-image-tools/cli_tools v0.0.0-20200504155228-251922788421
+	github.com/GoogleCloudPlatform/compute-image-tools/daisy v0.0.0-20200504155228-251922788421
+	github.com/GoogleCloudPlatform/compute-image-tools/gce_image_import_export_tests v0.0.0-20200117185841-203d88ed91bb
+	github.com/GoogleCloudPlatform/compute-image-tools/go/e2e_test_utils v0.0.0-20200504155228-251922788421
+	golang.org/x/net v0.0.0-20200501053045-e0ff5e5a1de5 // indirect
+	golang.org/x/sys v0.0.0-20200501145240-bc7a7d42d5c3 // indirect
+	golang.org/x/tools v0.0.0-20200504152539-33427f1b0364 // indirect
+	google.golang.org/api v0.22.0
+	google.golang.org/genproto v0.0.0-20200430143042-b979b6f78d84 // indirect
+)
