+++ cli_tools_e2e_test/gce_ovf_import_tests/main.go
@@ -23,9 +23,9 @@ import (
 
 	ovfinstanceimporttestsuite "github.com/GoogleCloudPlatform/compute-image-tools/gce_ovf_import_tests/test_suites/ovf_instance_import"
 	ovfmachineimageimporttestsuite "github.com/GoogleCloudPlatform/compute-image-tools/gce_ovf_import_tests/test_suites/ovf_machine_image_import"
-	"github.com/GoogleCloudPlatform/compute-image-tools/go/e2e_test_utils"
+	e2etestutils "github.com/GoogleCloudPlatform/compute-image-tools/go/e2e_test_utils"
 	"github.com/GoogleCloudPlatform/compute-image-tools/go/e2e_test_utils/junitxml"
-	"github.com/GoogleCloudPlatform/compute-image-tools/go/e2e_test_utils/test_config"
+	testconfig "github.com/GoogleCloudPlatform/compute-image-tools/go/e2e_test_utils/test_config"
 )
 
 func main() {
