+++ cli_tools_e2e_test/gce_ovf_import_tests/scripts/ovf_import_test_ubuntu_3_disks.sh
