+++ cli_tools_e2e_test/gce_ovf_import_tests/scripts/ovf_import_test_windows_two_disks.ps1
