+++ cli_tools_e2e_test/gce_ovf_import_tests/test_suites/ovf_instance_import/ovf_instance_import_test_suite.go
@@ -1,4 +1,4 @@
-//  Copyright 2020 Google Inc. All Rights Reserved.
+//  Copyright 2019 Google Inc. All Rights Reserved.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -12,9 +12,8 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-// Package ovfmachineimageimporttestsuite contains e2e tests for machine image
-// import cli tools
-package ovfmachineimageimporttestsuite
+// Package ovfinstanceimporttestsuite contains e2e tests for instance import cli tools
+package ovfinstanceimporttestsuite
 
 import (
 	"context"
@@ -33,11 +32,11 @@ import (
 	computeUtils "github.com/GoogleCloudPlatform/compute-image-tools/gce_ovf_import_tests/compute"
 	clitoolstestutils "github.com/GoogleCloudPlatform/compute-image-tools/go/e2e_test_utils/cli_tools"
 	"github.com/GoogleCloudPlatform/compute-image-tools/go/e2e_test_utils/junitxml"
-	"github.com/GoogleCloudPlatform/compute-image-tools/go/e2e_test_utils/test_config"
+	testconfig "github.com/GoogleCloudPlatform/compute-image-tools/go/e2e_test_utils/test_config"
 )
 
 const (
-	testSuiteName = "OVFMachineImageImportTests"
+	testSuiteName = "OVFInstanceImportTests"
 	ovaBucket     = "compute-image-tools-test-resources"
 )
 
@@ -49,8 +48,8 @@ var (
 	}
 )
 
-type ovfMachineImageImportTestProperties struct {
-	machineImageName          string
+type ovfInstanceImportTestProperties struct {
+	instanceName              string
 	isWindows                 bool
 	expectedStartupOutput     string
 	verificationStartupScript string
@@ -60,7 +59,6 @@ type ovfMachineImageImportTestProperties struct {
 	machineType               string
 	network                   string
 	subnet                    string
-	storageLocation           string
 }
 
 // TestSuite is image import test suite.
@@ -78,36 +76,51 @@ func TestSuite(
 		clitoolstestutils.GcloudLatestWrapperLatest,
 	}
 	for _, testType := range testTypes {
-		machineImageImportUbuntu3DisksTestCase := junitxml.NewTestCase(
-			testSuiteName, fmt.Sprintf("[%v][OVFMachineImageImport] %v", testType, "Ubuntu 3 disks, one data disk larger than 10GB"))
-		machineImageImportWindows2012R2TwoDisks := junitxml.NewTestCase(
-			testSuiteName, fmt.Sprintf("[%v][OVFMachineImageImport] %v", testType, "Windows 2012 R2 two disks"))
-		machineImageImportNetworkSettingsName := junitxml.NewTestCase(
-			testSuiteName, fmt.Sprintf("[%v][OVFMachineImageImport] %v", testType, "Network setting (name only)"))
-		machineImageImportNetworkSettingsPath := junitxml.NewTestCase(
-			testSuiteName, fmt.Sprintf("[%v][OVFMachineImageImport] %v", testType, "Network setting (path)"))
-		machineImageImportStorageLocation := junitxml.NewTestCase(
-			testSuiteName, fmt.Sprintf("[%v][OVFMachineImageImport] %v", testType, "Storage location"))
+		instanceImportUbuntu3DisksTestCase := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Ubuntu 3 disks, one data disk larger than 10GB"))
+		instanceImportCentos68 := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Centos 6.8"))
+		instanceImportWindows2012R2TwoDisks := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Windows 2012 R2 two disks"))
+		instanceImportWindows2016 := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Windows 2016"))
+		instanceImportWindows2008R2FourNICs := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Windows 2008r2 - Four NICs"))
+		instanceImportDebian9 := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Debian 9"))
+		instanceImportUbuntu16FromVirtualBox := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Ubuntu 1604 from Virtualbox"))
+		instanceImportUbuntu16FromAWS := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Ubuntu 1604 from AWS"))
+		instanceImportNetworkSettingsName := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Test network setting (name only)"))
+		instanceImportNetworkSettingsPath := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Test network setting (path)"))
 
 		testsMap[testType] = map[*junitxml.TestCase]func(
 			context.Context, *junitxml.TestCase, *log.Logger, *testconfig.Project, clitoolstestutils.CLITestType){}
-		testsMap[testType][machineImageImportUbuntu3DisksTestCase] = runOVFMachineImageImportUbuntu3Disks
-		testsMap[testType][machineImageImportWindows2012R2TwoDisks] = runOVFMachineImageImportWindows2012R2TwoDisks
-		testsMap[testType][machineImageImportNetworkSettingsName] = runOVFMachineImageImportNetworkSettingsName
-		testsMap[testType][machineImageImportNetworkSettingsPath] = runOVFMachineImageImportNetworkSettingsPath
-		testsMap[testType][machineImageImportStorageLocation] = runOVFMachineImageImportStorageLocation
+		testsMap[testType][instanceImportUbuntu3DisksTestCase] = runOVFInstanceImportUbuntu3Disks
+		testsMap[testType][instanceImportCentos68] = runOVFInstanceImportCentos68
+		testsMap[testType][instanceImportWindows2012R2TwoDisks] = runOVFInstanceImportWindows2012R2TwoDisks
+		testsMap[testType][instanceImportWindows2016] = runOVFInstanceImportWindows2016
+		testsMap[testType][instanceImportWindows2008R2FourNICs] = runOVFInstanceImportWindows2008R2FourNICs
+		testsMap[testType][instanceImportDebian9] = runOVFInstanceImportDebian9
+		testsMap[testType][instanceImportUbuntu16FromVirtualBox] = runOVFInstanceImportUbuntu16FromVirtualBox
+		testsMap[testType][instanceImportUbuntu16FromAWS] = runOVFInstanceImportUbuntu16FromAWS
+		testsMap[testType][instanceImportNetworkSettingsName] = runOVFInstanceImportNetworkSettingsName
+		testsMap[testType][instanceImportNetworkSettingsPath] = runOVFInstanceImportNetworkSettingsPath
 	}
 
 	clitoolstestutils.CLITestSuite(ctx, tswg, testSuites, logger, testSuiteRegex, testCaseRegex,
 		testProjectConfig, testSuiteName, testsMap)
 }
 
-func runOVFMachineImageImportUbuntu3Disks(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+func runOVFInstanceImportUbuntu3Disks(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
 	testProjectConfig *testconfig.Project, testType clitoolstestutils.CLITestType) {
 
 	suffix := path.RandString(5)
-	props := &ovfMachineImageImportTestProperties{
-		machineImageName: fmt.Sprintf("test-machine-image-ubuntu-3-disks-%v", suffix),
+	props := &ovfInstanceImportTestProperties{
+		instanceName: fmt.Sprintf("test-instance-ubuntu-3-disks-%v", suffix),
 		verificationStartupScript: loadScriptContent(
 			"gce_ovf_import_tests/scripts/ovf_import_test_ubuntu_3_disks.sh", logger),
 		zone:                  testProjectConfig.TestZone,
@@ -116,15 +129,15 @@ func runOVFMachineImageImportUbuntu3Disks(ctx context.Context, testCase *junitxm
 		os:                    "ubuntu-1604",
 		machineType:           "n1-standard-4"}
 
-	runOVFMachineImageImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
 }
 
-func runOVFMachineImageImportCentos68(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+func runOVFInstanceImportCentos68(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
 	testProjectConfig *testconfig.Project, testType clitoolstestutils.CLITestType) {
 
 	suffix := path.RandString(5)
-	props := &ovfMachineImageImportTestProperties{
-		machineImageName: fmt.Sprintf("test-machine-image-centos-6-%v", suffix),
+	props := &ovfInstanceImportTestProperties{
+		instanceName: fmt.Sprintf("test-instance-centos-6-%v", suffix),
 		verificationStartupScript: loadScriptContent(
 			"daisy_integration_tests/scripts/post_translate_test.sh", logger),
 		zone:                  testProjectConfig.TestZone,
@@ -134,15 +147,15 @@ func runOVFMachineImageImportCentos68(ctx context.Context, testCase *junitxml.Te
 		machineType:           "n1-standard-4",
 	}
 
-	runOVFMachineImageImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
 }
 
-func runOVFMachineImageImportWindows2012R2TwoDisks(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+func runOVFInstanceImportWindows2012R2TwoDisks(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
 	testProjectConfig *testconfig.Project, testType clitoolstestutils.CLITestType) {
 
 	suffix := path.RandString(5)
-	props := &ovfMachineImageImportTestProperties{
-		machineImageName: fmt.Sprintf("test-machine-image-w2k12-r2-%v", suffix),
+	props := &ovfInstanceImportTestProperties{
+		instanceName: fmt.Sprintf("test-instance-w2k12-r2-%v", suffix),
 		verificationStartupScript: loadScriptContent(
 			"gce_ovf_import_tests/scripts/ovf_import_test_windows_two_disks.ps1", logger),
 		zone:                  testProjectConfig.TestZone,
@@ -153,15 +166,15 @@ func runOVFMachineImageImportWindows2012R2TwoDisks(ctx context.Context, testCase
 		isWindows:             true,
 	}
 
-	runOVFMachineImageImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
 }
 
-func runOVFMachineImageImportStorageLocation(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+func runOVFInstanceImportWindows2016(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
 	testProjectConfig *testconfig.Project, testType clitoolstestutils.CLITestType) {
 
 	suffix := path.RandString(5)
-	props := &ovfMachineImageImportTestProperties{
-		machineImageName: fmt.Sprintf("test-machine-image-w2k16-%v", suffix),
+	props := &ovfInstanceImportTestProperties{
+		instanceName: fmt.Sprintf("test-instance-w2k16-%v", suffix),
 		verificationStartupScript: loadScriptContent(
 			"daisy_integration_tests/scripts/post_translate_test.ps1", logger),
 		zone:                  testProjectConfig.TestZone,
@@ -170,18 +183,85 @@ func runOVFMachineImageImportStorageLocation(ctx context.Context, testCase *juni
 		os:                    "windows-2016",
 		machineType:           "n2-standard-2",
 		isWindows:             true,
-		storageLocation:       "us-west2",
 	}
 
-	runOVFMachineImageImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
 }
 
-func runOVFMachineImageImportNetworkSettingsName(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+func runOVFInstanceImportWindows2008R2FourNICs(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
 	testProjectConfig *testconfig.Project, testType clitoolstestutils.CLITestType) {
 
 	suffix := path.RandString(5)
-	props := &ovfMachineImageImportTestProperties{
-		machineImageName: fmt.Sprintf("test-network-name-%v", suffix),
+	props := &ovfInstanceImportTestProperties{
+		instanceName: fmt.Sprintf("test-instance-w2k8r2-%v", suffix),
+		verificationStartupScript: loadScriptContent(
+			"daisy_integration_tests/scripts/post_translate_test.ps1", logger),
+		zone:                  testProjectConfig.TestZone,
+		expectedStartupOutput: "All Tests Passed",
+		sourceURI:             fmt.Sprintf("gs://%v/ova/win2008r2-all-updates-four-nic.ova", ovaBucket),
+		os:                    "windows-2008r2",
+		isWindows:             true,
+	}
+
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+}
+
+func runOVFInstanceImportDebian9(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+	testProjectConfig *testconfig.Project, testType clitoolstestutils.CLITestType) {
+
+	suffix := path.RandString(5)
+	// no startup script as this OVA has issues running it (possibly due to no SSH allowed)
+	// b/141321520
+	props := &ovfInstanceImportTestProperties{
+		instanceName: fmt.Sprintf("test-instance-debian-9-%v", suffix),
+		zone:         testProjectConfig.TestZone,
+		sourceURI:    fmt.Sprintf("gs://%v/ova/bitnami-tomcat-8.5.43-0-linux-debian-9-x86_64.ova", ovaBucket),
+		os:           "debian-9",
+		machineType:  "n1-standard-4",
+	}
+
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+}
+
+func runOVFInstanceImportUbuntu16FromVirtualBox(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+	testProjectConfig *testconfig.Project, testType clitoolstestutils.CLITestType) {
+
+	suffix := path.RandString(5)
+	props := &ovfInstanceImportTestProperties{
+		instanceName: fmt.Sprintf("test-instance-virtualbox-6-%v", suffix),
+		verificationStartupScript: loadScriptContent(
+			"daisy_integration_tests/scripts/post_translate_test.sh", logger),
+		zone:                  testProjectConfig.TestZone,
+		expectedStartupOutput: "All tests passed!",
+		sourceURI:             fmt.Sprintf("gs://%v/ova/ubuntu-16.04-virtualbox.ova", ovaBucket),
+		os:                    "ubuntu-1604",
+		machineType:           "n1-standard-4",
+	}
+
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+}
+
+func runOVFInstanceImportUbuntu16FromAWS(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+	testProjectConfig *testconfig.Project, testType clitoolstestutils.CLITestType) {
+
+	suffix := path.RandString(5)
+	props := &ovfInstanceImportTestProperties{
+		instanceName: fmt.Sprintf("test-instance-aws-ova-ubuntu-1604-%v", suffix),
+		zone:         testProjectConfig.TestZone,
+		sourceURI:    fmt.Sprintf("gs://%v/ova/aws-ova-ubuntu-1604.ova", ovaBucket),
+		os:           "ubuntu-1604",
+		machineType:  "n1-standard-4",
+	}
+
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+}
+
+func runOVFInstanceImportNetworkSettingsName(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+	testProjectConfig *testconfig.Project, testType clitoolstestutils.CLITestType) {
+
+	suffix := path.RandString(5)
+	props := &ovfInstanceImportTestProperties{
+		instanceName: fmt.Sprintf("test-network-name-%v", suffix),
 		verificationStartupScript: loadScriptContent(
 			"daisy_integration_tests/scripts/post_translate_test.sh", logger),
 		zone:                  testProjectConfig.TestZone,
@@ -193,16 +273,16 @@ func runOVFMachineImageImportNetworkSettingsName(ctx context.Context, testCase *
 		subnet:                fmt.Sprintf("%v-subnet-1", testProjectConfig.TestProjectID),
 	}
 
-	runOVFMachineImageImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
 }
 
-func runOVFMachineImageImportNetworkSettingsPath(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+func runOVFInstanceImportNetworkSettingsPath(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
 	testProjectConfig *testconfig.Project, testType clitoolstestutils.CLITestType) {
 
 	suffix := path.RandString(5)
 	region, _ := paramhelper.GetRegion(testProjectConfig.TestZone)
-	props := &ovfMachineImageImportTestProperties{
-		machineImageName: fmt.Sprintf("test-network-path-%v", suffix),
+	props := &ovfInstanceImportTestProperties{
+		instanceName: fmt.Sprintf("test-network-path-%v", suffix),
 		verificationStartupScript: loadScriptContent(
 			"daisy_integration_tests/scripts/post_translate_test.sh", logger),
 		zone:                  testProjectConfig.TestZone,
@@ -214,21 +294,22 @@ func runOVFMachineImageImportNetworkSettingsPath(ctx context.Context, testCase *
 		subnet:                fmt.Sprintf("projects/%v/regions/%v/subnetworks/%v-subnet-1", testProjectConfig.TestProjectID, region, testProjectConfig.TestProjectID),
 	}
 
-	runOVFMachineImageImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
 }
 
-func buildTestArgs(props *ovfMachineImageImportTestProperties, testProjectConfig *testconfig.Project) map[clitoolstestutils.CLITestType][]string {
+func buildTestArgs(props *ovfInstanceImportTestProperties, testProjectConfig *testconfig.Project) map[clitoolstestutils.CLITestType][]string {
 	gcloudArgs := []string{
-		"beta", "compute", "machine-images", "import", props.machineImageName, "--quiet",
+		"beta", "compute", "instances", "import", props.instanceName, "--quiet",
 		"--docker-image-tag=latest",
 		fmt.Sprintf("--project=%v", testProjectConfig.TestProjectID),
 		fmt.Sprintf("--source-uri=%v", props.sourceURI),
 		fmt.Sprintf("--zone=%v", testProjectConfig.TestZone),
 	}
 	wrapperArgs := []string{"-client-id=e2e", fmt.Sprintf("-project=%v", testProjectConfig.TestProjectID),
-		fmt.Sprintf("-machine-image-name=%s", props.machineImageName),
+		fmt.Sprintf("-instance-names=%s", props.instanceName),
 		fmt.Sprintf("-ovf-gcs-path=%v", props.sourceURI),
 		fmt.Sprintf("-zone=%v", testProjectConfig.TestZone),
+		fmt.Sprintf("-build-id=%v", path.RandString(10)),
 	}
 
 	if props.os != "" {
@@ -247,10 +328,6 @@ func buildTestArgs(props *ovfMachineImageImportTestProperties, testProjectConfig
 		gcloudArgs = append(gcloudArgs, fmt.Sprintf("--subnet=%v", props.subnet))
 		wrapperArgs = append(wrapperArgs, fmt.Sprintf("-subnet=%v", props.subnet))
 	}
-	if props.storageLocation != "" {
-		gcloudArgs = append(gcloudArgs, fmt.Sprintf("--storage-location=%v", props.storageLocation))
-		wrapperArgs = append(wrapperArgs, fmt.Sprintf("-machine-image-storage-location=%v", props.storageLocation))
-	}
 
 	argsMap := map[clitoolstestutils.CLITestType][]string{
 		clitoolstestutils.Wrapper:                   wrapperArgs,
@@ -260,12 +337,12 @@ func buildTestArgs(props *ovfMachineImageImportTestProperties, testProjectConfig
 	return argsMap
 }
 
-func runOVFMachineImageImportTest(ctx context.Context, args []string, testType clitoolstestutils.CLITestType,
+func runOVFInstanceImportTest(ctx context.Context, args []string, testType clitoolstestutils.CLITestType,
 	testProjectConfig *testconfig.Project, logger *log.Logger, testCase *junitxml.TestCase,
-	props *ovfMachineImageImportTestProperties) {
+	props *ovfInstanceImportTestProperties) {
 
 	if clitoolstestutils.RunTestForTestType(cmds[testType], args, testType, logger, testCase) {
-		verifyImportedMachineImage(ctx, testCase, testProjectConfig, logger, props)
+		verifyImportedInstance(ctx, testCase, testProjectConfig, logger, props)
 	}
 }
 
@@ -274,9 +351,9 @@ func failure(testCase *junitxml.TestCase, logger *log.Logger, msg string) {
 	logger.Printf(msg)
 }
 
-func verifyImportedMachineImage(
+func verifyImportedInstance(
 	ctx context.Context, testCase *junitxml.TestCase, testProjectConfig *testconfig.Project,
-	logger *log.Logger, props *ovfMachineImageImportTestProperties) {
+	logger *log.Logger, props *ovfInstanceImportTestProperties) {
 
 	client, err := daisyCompute.NewClient(ctx)
 	if err != nil {
@@ -284,29 +361,19 @@ func verifyImportedMachineImage(
 		return
 	}
 
-	logger.Printf("Verifying imported machine image...")
-	testInstanceName := props.machineImageName + "-test-instance"
-	logger.Printf("Creating `%v` test instance.", testInstanceName)
-	instance, err := computeUtils.CreateInstanceBeta(
-		ctx, testProjectConfig.TestProjectID, props.zone, testInstanceName, props.isWindows, props.machineImageName)
+	logger.Printf("Verifying imported instance...")
+	instance, err := computeUtils.CreateInstanceObject(ctx, testProjectConfig.TestProjectID, props.zone, props.instanceName, props.isWindows)
 	if err != nil {
-		failure(testCase, logger, fmt.Sprintf("Error when creating test instance `%v` from machine image '%v': %v", testInstanceName, props.machineImageName, err))
+		failure(testCase, logger, fmt.Sprintf("Image '%v' doesn't exist after import: %v", props.instanceName, err))
 		return
 	}
 
-	// Clean-up
 	defer func() {
-		logger.Printf("Deleting instance `%v`", testInstanceName)
+		logger.Printf("Deleting instance `%v`", props.instanceName)
 		if err := instance.Cleanup(); err != nil {
-			logger.Printf("Instance '%v' failed to clean up: %v", testInstanceName, err)
+			logger.Printf("Instance '%v' failed to clean up: %v", props.instanceName, err)
 		} else {
-			logger.Printf("Instance '%v' cleaned up.", testInstanceName)
-		}
-		logger.Printf("Deleting machine image `%v`", props.machineImageName)
-		if err := client.DeleteMachineImage(testProjectConfig.TestProjectID, props.machineImageName); err != nil {
-			logger.Printf("Machine image '%v' failed to clean up: %v", props.machineImageName, err)
-		} else {
-			logger.Printf("Machine image '%v' cleaned up.", props.machineImageName)
+			logger.Printf("Instance '%v' cleaned up.", props.instanceName)
 		}
 	}()
 
@@ -349,17 +416,26 @@ func verifyImportedMachineImage(
 		return
 	}
 
+	logger.Printf("[%v] Stopping instance before restarting with test startup script", props.instanceName)
+	err = client.StopInstance(
+		testProjectConfig.TestProjectID, props.zone, props.instanceName)
+
+	if err != nil {
+		testCase.WriteFailure("Error stopping imported instance: %v", err)
+		return
+	}
+
 	if props.verificationStartupScript == "" {
-		logger.Printf("[%v] Will not set test startup script to test instance metadata as it's not defined", props.machineImageName)
+		logger.Printf("[%v] Will not set test startup script to instance metadata as it's not defined", props.instanceName)
 		return
 	}
 
 	err = instance.StartWithScript(props.verificationStartupScript)
 	if err != nil {
-		testCase.WriteFailure("Error starting instance `%v` with script: `%v`: %v", testInstanceName, err)
+		testCase.WriteFailure("Error starting instance `%v` with script: %v", props.instanceName, err)
 		return
 	}
-	logger.Printf("[%v] Waiting for `%v` in instance serial console.", testInstanceName,
+	logger.Printf("[%v] Waiting for `%v` in instance serial console.", props.instanceName,
 		props.expectedStartupOutput)
 	if err := instance.WaitForSerialOutput(
 		props.expectedStartupOutput, 1, 5*time.Second, 15*time.Minute); err != nil {
