--- gce_image_import_export_tests/go.mod
@@ -1,11 +1,0 @@
-module github.com/GoogleCloudPlatform/compute-image-tools/gce_image_import_export_tests
-
-go 1.13
-
-require (
-	cloud.google.com/go/storage v1.5.0
-	github.com/GoogleCloudPlatform/compute-image-tools/cli_tools v0.0.0-20200114193945-760dc9647411
-	github.com/GoogleCloudPlatform/compute-image-tools/daisy v0.0.0-20200114193104-06c8be9a6a7d
-	github.com/GoogleCloudPlatform/compute-image-tools/go/e2e_test_utils v0.0.0-20200114194943-8842c1d85084
-	google.golang.org/api v0.15.0
-)
