--- gce_ovf_import_tests/go.mod
@@ -1,12 +1,0 @@
-module github.com/GoogleCloudPlatform/compute-image-tools/gce_ovf_import_tests
-
-go 1.13
-
-require (
-	github.com/GoogleCloudPlatform/compute-image-tools/cli_tools v0.0.0-20200206200800-33f55f710a20
-	github.com/GoogleCloudPlatform/compute-image-tools/daisy v0.0.0-20200206200800-33f55f710a20
-	github.com/GoogleCloudPlatform/compute-image-tools/gce_image_import_export_tests v0.0.0-20200117185841-203d88ed91bb
-	github.com/GoogleCloudPlatform/compute-image-tools/go/e2e_test_utils v0.0.0-20200206200800-33f55f710a20
-	golang.org/x/tools v0.0.0-20200206204726-37215997d4fb // indirect
-	google.golang.org/api v0.17.0
-)
